refactor: use shared folder for attriibute quard
