+++ src/api/types/guards.ts
@@ -1,0 +1,8 @@
+export const isAttribute = (value: unknown): value is { key: string; label: string } => {
+  return (
+    value != null &&
+    typeof value === 'object' &&
+    Reflect.has(value, 'key') &&
+    Reflect.has(value, 'label')
+  );
+};
